refactor: split lib/tasks.ts into modular components (Phase 3, Task 1)
Split lib/tasks.ts from 555 lines into focused modules:
- lib/tasks/crud.ts (295 lines) - Core CRUD operations
- lib/tasks/subtasks.ts (119 lines) - Subtask management
- lib/tasks/dependencies.ts (100 lines) - Dependency operations
- lib/tasks/import-export.ts (89 lines) - Import/export logic
- lib/tasks.ts (45 lines) - Re-export layer for backward compatibility

Benefits:
- All files now under 300-line coding standard limit
- Single responsibility per module
- Improved maintainability and testability
- Zero breaking changes (100% backward compatible)

Verification:
- TypeScript compilation: No new errors
- Test suite: 479/481 tests passing (2 pre-existing failures)
- All existing imports from @/lib/tasks work unchanged

Related: Phase 3 of coding standards compliance (see next.md)
